+++ packages/core/src/index.ts
@@ -2,3 +2,9 @@ export * from './lib/sdk.js';
 export * from './lib/types/sdk.config.js';
 
 export * from './lib/types/store/index.js';
+
+// Common
+export * from './lib/types/sdk.config.js';
+export * from './lib/types/api.js';
+export * from './lib/plugins/index.js';
+export * from './lib/utilities/axios.utility.js';