+++ packages/core/src/lib/services/store/product.attribute.service.ts
@@ -1,6 +1,7 @@
 import { ProductAttributeResponse } from '../../types/store/product-attribute/product.attribute.response.js';
 import { ApiResult } from '../../types/api.js';
 import { BaseService } from '../base.service.js';
+import { doGet } from '../../utilities/axios.utility.js';
 
 /**
  * Product Attributes API
@@ -14,7 +15,7 @@ export class ProductAttributeService extends BaseService {
    */
   async list(): Promise<ApiResult<ProductAttributeResponse[]>> {
     const url = `${this.baseUrl}/${this.endpoint}`;
-    const { data, error } = await this.doGet<ProductAttributeResponse[]>(url);
+    const { data, error } = await doGet<ProductAttributeResponse[]>(url);
     return { data, error };
   }
 
@@ -25,7 +26,7 @@ export class ProductAttributeService extends BaseService {
    */
   async single(id: number): Promise<ApiResult<ProductAttributeResponse>> {
     const url = `${this.baseUrl}/${this.endpoint}/${id}`;
-    const { data, error } = await this.doGet<ProductAttributeResponse>(url);
+    const { data, error } = await doGet<ProductAttributeResponse>(url);
     return { data, error };
   }
 }