+++ packages/core/src/lib/services/store/product.category.service.ts
@@ -3,6 +3,7 @@ import { ProductCategoryRequest } from '../../types/store/product-category/produ
 import qs from 'qs';
 import { ApiResult } from '../../types/api.js';
 import { BaseService } from '../base.service.js';
+import { doGet } from '../../utilities/axios.utility.js';
 
 /**
  * Product Categories API
@@ -20,7 +21,7 @@ export class ProductCategoryService extends BaseService {
   ): Promise<ApiResult<ProductCategoryResponse[]>> {
     const query = qs.stringify(params);
     const url = `${this.baseUrl}/${this.endpoint}?${query}`;
-    const { data, error } = await this.doGet<ProductCategoryResponse[]>(url);
+    const { data, error } = await doGet<ProductCategoryResponse[]>(url);
     return { data, error };
   }
 
@@ -31,7 +32,7 @@ export class ProductCategoryService extends BaseService {
    */
   async single(id: number): Promise<ApiResult<ProductCategoryResponse>> {
     const url = `${this.baseUrl}/${this.endpoint}/${id}`;
-    const { data, error } = await this.doGet<ProductCategoryResponse>(url);
+    const { data, error } = await doGet<ProductCategoryResponse>(url);
     return { data, error };
   }
 }