+++ packages/hippoo/src/lib/types/refresh-token/refresh.token.response.ts
@@ -1,0 +1,9 @@
+export interface RefreshTokenResponse {
+  token: string;
+  refresh_token: string;
+  user: {
+    id: number;
+    name: string;
+    email: string;
+  };
+}