+++ packages/hippoo/src/lib/types/signup/signup.response.ts
@@ -1,0 +1,9 @@
+export interface SignupResponse {
+  token: string;
+  refresh_token: string;
+  user: {
+    id: number;
+    email: string;
+    name: string;
+  };
+}